Update of the Static craze
It's in the name.. I went on a bit of a static frenzy. A shit ton of
static methods were added. It's really helpful to not have to make a
boolean that's equal to a boolean in a hashmap etc etc

List of Static methods:
-addStock, adds (or minus) stock to a player. It's helpful.
-setStock, sets the stock of a player
-getStock, returns the stock of a player

-addKnockback, adds knockback to a player
-setKnockback, sets the knockback of a player
-getKnockback, returns the knockback of a player

-setGame, sets the game state
-getGame, returns the game state

I also have some static methods for other things, but they're only used
in one class, so it's not worth mentioning.

It's worth nothing that all of the changes made in the .java files are
replacing HashMap methods with the new static methods.

Respawns got a bit of an overhaul. They work relatively the same though.

-CooldownManager started